requirements.txt
Django==2.2
django-ckeditor==5.7.1
Pillow==6.1.0
django-bootstrap-form==3.4
django-pagination-bootstrap==1.3